+++ src/main/java/it/univpm/utils/ParserPressione.java
@@ -1,7 +1,6 @@
 package it.univpm.utils;
 
 import it.univpm.models.Pressione;
-import it.univpm.models.Temperatura;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.ParseException;
@@ -26,9 +25,9 @@ public class ParserPressione extends Parser{
             JSONObject obj = (JSONObject) listaAPI.get(i);
             long dataEpoch = (long) obj.get("dt");
             JSONObject main = (JSONObject) obj.get("main");
-            int valorePressione = (int) main.get("pressure");
-            int valoreMare = (int) main.get("sea_level");
-            int valoreSuolo = (int) main.get("grnd_level");
+            long valorePressione = (long) main.get("pressure");
+            long valoreMare = (long) main.get("sea_level");
+            long valoreSuolo = (long) main.get("grnd_level");
             Pressione nuovaPress = new Pressione(valorePressione, valoreMare, valoreSuolo, dataEpoch);
             pressioni.add(nuovaPress);
         }
